% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz_centsums.R
\name{centcosums-accessor}
\alias{centcosums-accessor}
\alias{cosums}
\alias{cosums,centcosums-method}
\alias{sums,centcosums-method}
\alias{comoments}
\alias{comoments,centcosums-method}
\title{Accessor methods.}
\usage{
cosums(x)

\S4method{cosums}{centcosums}(x)

comoments(x, type = c("central", "raw"), used_df = 0)

\S4method{comoments}{centcosums}(x, type = c("central", "raw"), used_df = 0)
}
\arguments{
\item{x}{a \code{centcosums} object.}

\item{type}{the type of moment to compute.}

\item{used_df}{the number of used degrees of freedom, only for computing
the covariance, and only for the \sQuote{central} type of comoment.}
}
\description{
Access slot data from a \code{centcosums} object.
}
\note{
The moment computations provided by fromo are 
numerically robust, but will often \emph{not} provide the
same results as the 'standard' implementations,
due to differences in roundoff. We make every attempt to balance
speed and robustness. User assumes all risk from using
the fromo package.
}
\author{
Steven E. Pav \email{shabbychef@gmail.com}
}
