% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz_centsums.R
\docType{class}
\name{centcosums-class}
\alias{centcosums-class}
\alias{initialize,centcosums-method}
\alias{initialize,centcosums-class}
\alias{centcosums}
\title{centcosums Class.}
\usage{
\S4method{initialize}{centcosums}(.Object, cosums, order = NA_real_)

centcosums(cosums, order = NULL)
}
\arguments{
\item{.Object}{a \code{centcosums} object, or proto-object.}

\item{cosums}{the output of \code{\link{cent_cosums}}, say.}

\item{order}{the order, defaulting to \code{2}.}
}
\value{
An object of class \code{centcosums}.
}
\description{
An S4 class to store (centered) cosums of data, and to support operations on 
the same.
A cosum is the sum of products of different elements of the vector input.
}
\details{
A \code{centcosums} object contains a multidimensional array (now only
2-dimensional), as output by \code{cent_cosums}.
}
\section{Slots}{

\describe{
\item{\code{cosums}}{a multidimensional array of the cosums.
might be the output of \code{cent_cosums}, for example.}

\item{\code{order}}{the maximum order. ignored for now.
only a value of 2 is understood for now.}
}}

\note{
The moment computations provided by fromo are 
numerically robust, but will often \emph{not} provide the
same results as the 'standard' implementations,
due to differences in roundoff. We make every attempt to balance
speed and robustness. User assumes all risk from using
the fromo package.
}
\examples{
set.seed(123)
x <- matrix(rnorm(100*3),ncol=3)
obj <- new("centcosums",cosums=cent_cosums(x,max_order=2),order=2)
# equivalently
obj2 <- centcosums(cent_cosums(x, max_order=2),order=2)

}
\references{
Terriberry, T. "Computing Higher-Order Moments Online."
\url{https://web.archive.org/web/20140423031833/http://people.xiph.org/~tterribe/notes/homs.html}

J. Bennett, et. al., "Numerically Stable, Single-Pass, 
Parallel Statistics Algorithms," Proceedings of IEEE
International Conference on Cluster Computing, 2009.
\doi{10.1109/CLUSTR.2009.5289161}

Cook, J. D. "Accurately computing running variance."
\url{https://www.johndcook.com/standard_deviation/}

Cook, J. D. "Comparing three methods of computing 
standard deviation."
\url{https://www.johndcook.com/blog/2008/09/26/comparing-three-methods-of-computing-standard-deviation/}
}
\seealso{
cent_cosums
}
\author{
Steven E. Pav \email{shabbychef@gmail.com}
}
\keyword{moments}
