% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz_centsums.R
\name{c.centcosums}
\alias{c.centcosums}
\title{concatenate centcosums objects.}
\usage{
\\method{c}{centcosums}(...)
}
\arguments{
\item{...}{\code{centcosums} objects}
}
\description{
Concatenate centcosums objects.
}
\seealso{
join_cent_cosums
}

