% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz_centsums.R
\docType{methods}
\name{\%-\%,centcosums,centcosums-method}
\alias{\%-\%,centcosums,centcosums-method}
\title{unconcatenate centcosums objects.}
\usage{
\S4method{\%-\%}{centcosums,centcosums}(x, y)
}
\arguments{
\item{x}{a \code{centcosums} objects}

\item{y}{a \code{centcosums} objects}
}
\description{
Unconcatenate centcosums objects.
}
\seealso{
unjoin_cent_cosums
}
