% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz_centsums.R
\name{accessor}
\alias{accessor}
\alias{sums}
\alias{sums,centsums-method}
\alias{moments}
\alias{moments,centsums-method}
\title{Accessor methods.}
\usage{
sums(x)

\S4method{sums}{centsums}(x)

moments(x, type = c("central", "raw", "standardized"))

\S4method{moments}{centsums}(x, type = c("central", "raw", "standardized"))
}
\arguments{
\item{x}{a \code{centsums} object.}

\item{type}{the type of moment to compute.}
}
\value{
for raw type, returns all the raw moments. for central and
standardized, drops the mean, and returns the central or standardized 2nd
through kth moments.
}
\description{
Access slot data from a \code{centsums} object.
}
\note{
The moment computations provided by fromo are 
numerically robust, but will often \emph{not} provide the
same results as the 'standard' implementations,
due to differences in roundoff. We make every attempt to balance
speed and robustness. User assumes all risk from using
the fromo package.

Currently \sQuote{raw} moments returns mean and higher order
moments, while \sQuote{central} and \sQuote{standardized} drop the mean,
thus the size of the output depends on the \code{type}. This feels wrong to
me and may be subject to change in upcoming versions.
}
\author{
Steven E. Pav \email{shabbychef@gmail.com}
}
