% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz_centsums.R
\name{c.centsums}
\alias{c.centsums}
\title{concatenate centsums objects.}
\usage{
\\method{c}{centsums}(...)
}
\arguments{
\item{...}{\code{centsums} objects}
}
\description{
Concatenate centsums objects.
}
\seealso{
join_cent_sums
}
