% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz_centsums.R
\docType{methods}
\name{\%-\%}
\alias{\%-\%}
\alias{\%-\%,centsums,centsums-method}
\title{unconcatenate centsums objects.}
\usage{
x \%-\% y

\S4method{\%-\%}{centsums,centsums}(x, y)
}
\arguments{
\item{x}{a \code{centsums} objects}

\item{y}{a \code{centsums} objects}
}
\description{
Unconcatenate centsums objects.
}
\seealso{
unjoin_cent_sums
}

