% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{cent2raw}
\alias{cent2raw}
\title{Convert between different types of moments, raw, central, standardized.}
\usage{
cent2raw(input)
}
\arguments{
\item{input}{a vector of the count, then the mean, then the \code{2} through \code{k}
raw or central moments.}
}
\description{
Given raw or central or standardized moments, convert to another type.
}
\note{
The moment computations provided by fromo are 
numerically robust, but will often \emph{not} provide the
same results as the 'standard' implementations,
due to differences in roundoff. We make every attempt to balance
speed and robustness. User assumes all risk from using
the fromo package.
}
\author{
Steven E. Pav \email{shabbychef@gmail.com}
}
