% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fromo.r
\name{fromo-NEWS}
\alias{fromo-NEWS}
\title{News for package 'fromo':}
\description{
News for package 'fromo'

\newcommand{\CRANpkg}{\href{http://CRAN.R-project.org/package=#1}{\pkg{#1}}}
\newcommand{\cranfromo}{\CRANpkg{fromo}}
\newcommand{\fromo}{\href{https://github.com/shabbychef/fromo}}
}
\section{\fromo{} Version 0.1.3 (2016-04-04) }{

\itemize{
\item submit to CRAN
}
}

\section{\fromo{} Initial Version 0.1.0 (2016-03-25) }{

\itemize{
\item start work
}
}

