% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fromo.r
\name{fromo-NEWS}
\alias{fromo-NEWS}
\title{News for package 'fromo':}
\description{
News for package 'fromo'

\newcommand{\CRANpkg}{\href{https://cran.r-project.org/package=#1}{\pkg{#1}}}
\newcommand{\cranfromo}{\CRANpkg{fromo}}
\newcommand{\fromo}{\href{https://github.com/shabbychef/fromo}}
}
\section{\fromo{} Version 0.2.3 (2024-11-07) }{

\itemize{
\item adding flag and checking for negative even moments due to roundoff.
\item fixing divide by zero which would corrupt results when input was constant.
}
}

\section{\fromo{} Version 0.2.2 (2024-11-03) }{

\itemize{
\item fix \code{t_running_sum} and others to act as documented when
\code{variable_win} is flagged.
}
}

\section{\fromo{} Version 0.2.1 (2019-01-29) }{

\itemize{
\item fix memory leak for case where the mean only need be computed via a Welford object.
}
}

\section{\fromo{} Version 0.2.0 (2019-01-12) }{

\itemize{
\item add \code{std_cumulants}
\item add \code{\link{running_sum}}, \code{\link{running_mean}}.
\item Kahan compensated summation for these.
\item Welford object under the hood.
\item add weighted moments computation.
\item add time-based running window computations.
\item some speedups for obviously fast cases: no checking of NA, etc.
\item move github figures to location CRAN understands.
}
}

\section{\fromo{} Version 0.1.3 (2016-04-04) }{

\itemize{
\item submit to CRAN
}
}

\section{\fromo{} Initial Version 0.1.0 (2016-03-25) }{

\itemize{
\item start work
}
}

